refactor: Controllers and models billed to be cleaner

+++ src/main/java/emannoel/finance/service/bank/DeleteBankingService.java
@@ -1,19 +1,29 @@
 package emannoel.finance.service.bank;
 
+import emannoel.finance.exceptions.NotFoundException;
+import emannoel.finance.model.bank.Banking;
 import emannoel.finance.repository.BankingRepository;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class DeleteBankingService {
 
-    private BankingRepository bankingRepository;
+    private final BankingRepository bankingRepository;
     public DeleteBankingService(BankingRepository bankingRepository){
         this.bankingRepository = bankingRepository;
     }
 
-    public String handler(Long id){
-        this.bankingRepository.deleteById(id);
-        return "Banking Deleted";
+    public void handler(Long id) throws NotFoundException{
+
+        Optional<Banking> bank = this.bankingRepository.findById(id);
+        if(bank.isPresent()){
+            this.bankingRepository.deleteById(id);
+            return;
+        }
+        throw new NotFoundException("Bank Not found");
+
     }
 
 }
